requirements.txt
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.20.0
requests>=2.27.1
